requirements.txt
fastapi==0.127.0
uvicorn==0.40.0
python-dotenv==1.2.1
pydantic==2.12.5
supabase==1.0.2
